+++ spec/survey_monkey_tools_spec.rb
@@ -14,25 +14,42 @@ RSpec.describe SurveyMonkeyTools do
     end
   end
 
-  describe "#get" do
+  describe "#surveys" do
     context "when successfully accessed" do
       before do
         filepath = "fixtures/surveymonkey/get_surveys_response.json"
-        body = File.read(File.join(__dir__, filepath))
-        surveymonkey_request("surveys").to_return(body: body, status: 200)
+        @body = File.read(File.join(__dir__, filepath))
+        surveymonkey_request(SurveyMonkeyTools::END_POINTS[:surveys]).to_return(body: @body, status: 201)
       end
 
-      it "returns a success response" do
+      it "returns a correct stdout" do
         survey_monkey_tools = SurveyMonkeyTools::CLI.new
-        response = survey_monkey_tools.response_get_surveys
 
-        expect(response.code.to_i).to eq(200)
+        expect { survey_monkey_tools.surveys }.to output(@body).to_stdout
       end
     end
   end
 
+  describe "#folders" do
+    before do
+      filepath = "fixtures/surveymonkey/get_folders_response.json"
+      @body = File.read(File.join(__dir__, filepath))
+      surveymonkey_request(SurveyMonkeyTools::END_POINTS[:folders]).to_return(body: @body, status: 200)
+    end
+
+    it "returns a correct stdout" do
+      survey_monkey_tools = SurveyMonkeyTools::CLI.new
+
+      expect { survey_monkey_tools.folders }.to output(@body).to_stdout
+    end
+  end
+
+  describe "#create_folder" do
+    it "returns a success response"
+  end
+
   def surveymonkey_request(endpoint)
-    stub_request(:get, "#{SurveyMonkeyTools::BASE_URI}/v3/#{endpoint}").with(
+    stub_request(:get, "#{SurveyMonkeyTools::BASE_URI}/v3#{endpoint}").with(
       headers: {
         "Accept" => "application/json",
         "Authorization" => "Bearer #{ENV["ACCESS_TOKEN"]}"