+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.3.0] - 2023-08-30
+
+- Added SurveyMonkeyTools::CLI#folders
+- Changed SurveyMonkeyTools::CLI#get to SurveyMonkeyTools::CLI#surveys
+
 ## [0.2.1] - 2023-08-30
 
 - Update API doc